Fix touchpad issue introduced by not being more selective with _OSI
Turns out that we won't hang as long as we set acpi_backlight=vendor or
if acpi_override_rev=5. The thing about both of these is that we're no
longer setting acpi_osi=!, which means that we report yes for many
different _OSI calls... It appears that the magic combination for the
Dell BIOS involves saying yes only to _OSI("Linux") and setting
acpi_backlight=vendor. This lets us have our backlight keys reported to
the OS, and it makes the touchpad work as we expect.